refactor(Runtime): Cleanup "app_view_utils", hide the rootInjector.
I want to get rid of "AppViewUtils" at the end of this refactor, so start removing the top-level functions into other runtime support libraries. In addition, make it less attractive to try and import from "lib/src" by making the rootInjector inaccessible directly.

PiperOrigin-RevId: 239681091
