+++ _goldens/test/_files/change_detection.template.golden
@@ -17,6 +17,7 @@ import 'package:angular/angular.dart';
 import 'package:angular/src/runtime/dom_helpers.dart' as import8;
 import 'package:angular/src/core/linker/app_view_utils.dart' as import9;
 import 'package:angular/src/runtime/text_binding.dart' as import10;
+import 'package:angular/src/runtime/interpolate.dart' as import11;
 
 final List<dynamic> styles$CheckOnceComponent = const [];
 
@@ -535,7 +536,7 @@ class ViewOnPushChildComponent0 extends AppView<import1.OnPushChildComponent> {
   @override
   void detectChangesInternal() {
     final _ctx = ctx;
-    _textBinding_2.updateText(import9.interpolateString0(_ctx.name));
+    _textBinding_2.updateText(import11.interpolateString0(_ctx.name));
   }
 
   @override
