+++ _goldens/test/_files/directives/intl_component.template.golden
@@ -16,7 +16,7 @@ import 'dart:html' as import6;
 import 'package:angular/src/runtime.dart' as import7;
 import 'package:angular/angular.dart';
 import 'package:angular/src/runtime/dom_helpers.dart' as import9;
-import 'package:angular/src/core/linker/app_view_utils.dart' as import10;
+import 'package:angular/src/runtime/interpolate.dart' as import10;
 
 final List<dynamic> styles$CompWithoutFinal = const [];
