+++ _goldens/test/_files/inlined_ngif.template.golden
@@ -18,7 +18,7 @@ import 'package:angular/src/runtime.dart' as import8;
 import 'package:angular/angular.dart';
 import 'package:angular/src/runtime/dom_helpers.dart' as import10;
 import 'package:angular/src/core/linker/template_ref.dart';
-import 'package:angular/src/core/linker/app_view_utils.dart' as import12;
+import 'package:angular/src/runtime/interpolate.dart' as import12;
 
 final List<dynamic> styles$InlinedNgIfComponent = const [];
